Updates to try and get android working
